# Makefile: do more caching in make test-only