@Raynos 2017-04-18
So on travis its apparently 20s vs 13s so its 40% faster.